Remove validate.rs
This deletes the code that wasn't used, and moves the other code
into the specific modules where it is used.
