+++ fea-rs/src/lib.rs
@@ -3,11 +3,9 @@
 mod ast;
 mod parse;
 mod types;
-mod validate;
 
 pub use ast::{AstSink, Node, NodeOrToken};
 pub use parse::grammar::root;
 pub use parse::util;
 pub use parse::{DebugSink, Kind, Parser, SyntaxError, TokenSet};
-pub use validate::validate;
 pub use types::{GlyphMap, GlyphName};
